# cheang-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cheang API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Cheang Handyman Platform",
    "contact": {
      "name": "Support Desk",
      "email": "support@cheang.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Access token stored in cookies"
      }
    }
  },
  "paths": {
...
    "/api/v1/services/create": {
      "post": {
        "summary": "Create a new service listing",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "price",
                  "userRef"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Service name",
                    "example": "House Cleaning"
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed description of service",
                    "example": "Deep cleaning for apartments and residential homes."
                  },
                  "price": {
                    "type": "number",
                    "description": "Flat or hourly price",
                    "example": 35
                  },
                  "userRef": {
                    "type": "string",
                    "description": "Handyman/User reference ID",
                    "example": "60b8d2f5f1d2b80015c92842"
                  },
                  "image": {
                    "type": "string",
                    "description": "URL to service listing image",
                    "example": "https://example.com/cleaning.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service listing created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/services/delete/{id}": {
      "delete": {
        "summary": "Delete a service listing by ID",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
    "/api/v1/services/update/{id}": {
      "post": {
        "summary": "Update an existing service listing by ID",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Premium Cleaning Service"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated deep cleaning details..."
                  },
                  "price": {
                    "type": "number",
                    "example": 45
                  },
                  "image": {
                    "type": "string",
                    "example": "https://example.com/premium-cleaning.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service listing updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
    "/api/v1/services/get/{id}": {
      "get": {
        "summary": "Retrieve a service listing by ID",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Service listing details retrieved successfully"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
    "/api/v1/services/user/{id}": {
      "get": {
        "summary": "Get all service listings for a specific user ID",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID (Handyman ID)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user services retrieved successfully"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "summary": "Retrieve notifications for the logged in user",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of notifications per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/notifications/mark-read": {
      "put": {
        "summary": "Mark all notifications as read for the logged in user",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All notifications marked as read successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/notifications/{id}": {
      "delete": {
        "summary": "Soft delete a specific notification",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted successfully"
          },
          "401": {
            "description": "Unauthorized (if trying to delete another user's notification)"
          },
          "404": {
            "description": "Notification not found"
          }
        }
      }
    },
...
  },
...
}